Use GCC's packed attribute for protocol structs
Because we are just overlaying the protocol buffer
with our structs, they have to be guarded against the compiler
optimizing the structs with padding.

The attribute packed prevents this, but probably causes
worse performance and is unportable e.g. for systems that do not allow
unaligned memory access.

A better solution would be to serialize the byte stream into our protocol
structs - this way the compiler is free to optimize the structs.

http://programmers.stackexchange.com/questions/307223/is-there-a-standard-way-or-standard-alternative-to-packing-a-struct-in-c
http://www.catb.org/esr/structure-packing/#_other_packing_techniques