Forward declare stuff
Oh, so it is actually called 'forward declaration'.. Didn't remember that.

The point is, if you have header files referencing e.g. structs from each other,
they need to be 'forward declared' since it is kind of a chicken and egg problem:

There's no way to know the exact type of the struct at the only pass the C compiler makes,
thus it can be just left declared without details.