Take care of network byte order ASAP
So when we receive a packet, it should be converted into
host-byte order ASAP. This eases up debugging, and as soon as you
start to fiddle around with the values, you anyway need to have it converted.

Vice versa, when outputting, the about last things to do would be to convert
the fields to network byte order.