Properly(?) handle thread cancellation
The problem is, that the threads can make blocking system calls.

Multitude of schemes can be implemented to handle this, but I decided
to just pthread_cancel the threads from a separate signal processor.

Not sure if this has the potential to leak OS resources (e.g. if
the signal handler logic is non-reentrant)