Start as root but change to uid `nobody` before starting stack
This change is introduced because suddenly CAP_NET_ADMIN
stopped working for my Arch Linux. In essence, doing

$ sudo setcap cap_net_admin=ep $(which ip)

Has no effect anymore:

$ ip link set dev tap0 up
RTNETLINK answers: Operation not permitted

Searching the Internet, one error case is if the directory is mounted
with `nosuid`, but this does not seem to be the case in my Linux.